BRIA-RMBG-2.0-Video/requirements.txt
torch
accelerate
opencv-python
spaces
pillow
numpy
timm
kornia
prettytable
typing
scikit-image
#huggingface_hub
transformers>=4.39.1
gradio==4.44.1
gradio_imageslider
loadimg>=0.1.1
httpx[socks]
huggingface_hub==0.25.0
moviepy